.history/requirements_20260325213652.txt
pyinstaller>=6.0.0
pillow>=10.0.0
fastapi[standard]
uvicorn>=0.21.1
sqlmodel>=0.0.37
sqlalchemy>=2.0
python-jose[cryptography]>=3.5.0
requests>=2.31.0
python-multipart>=0.0.6
argon2-cffi==25.1.0
psycopg2-binary>=2.9.11
yt-dlp>=2026.3.17